WebApps/API Server Django/microsite/requirements.txt
absl-py==1.3.0
asgiref==3.6.0
astunparse==1.6.3
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
flatbuffers==23.1.4
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
joblib==1.3.2
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.1
mssql-django==1.3
nltk==3.8.1
numpy==1.24.1
oauthlib==3.2.2
openapi-codec==1.3.2
opt-einsum==3.3.0
packaging==23.0
pandas==2.1.3
protobuf==3.19.6
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyodbc==4.0.39
python-dateutil==2.8.2
pytz==2022.7
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.11.3
simplejson==3.18.1
six==1.16.0
sqlparse==0.4.3
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
